requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.11.0
apscheduler==3.11.0
attrs==25.3.0
audioread==3.0.1
av==14.4.0
backoff==2.2.1
boto3==1.35.60
botocore==1.35.99
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
cython==3.0.12
dataclasses-json==0.6.7
decorator==4.3.0
deepgram-sdk==4.8.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
fastapi==0.110.3
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
google-auth==2.40.3
google-crc32c==1.7.1
google-genai==1.39.1
gputil==1.4.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
inference-sdk==0.56.0
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.9
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.45.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
mypy-extensions==1.1.0
networkx==3.5
numba==0.62.0
numpy==2.0.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnxruntime==1.19.2
openai==1.109.1
opencv-contrib-python==4.12.0.88
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
piexif==1.1.3
pillow==11.0
platformdirs==4.4.0
pluggy==1.6.0
pooch==1.8.2
proglog==0.1.12
prometheus-client==0.23.1
prometheus-fastapi-instrumentator==6.0.0
propcache==0.3.2
protobuf==6.32.1
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.0.2
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.11.0
pydot==2.0.0
pyee==13.0.0
pygments==2.19.2
pylibsrtp==0.12.0
pyopenssl==25.3.0
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
redis==5.0.8
regex==2025.9.18
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.0.1
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.6.2
scikit-image==0.24.0
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.37.2
structlog==24.4.0
supervision==0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.9.20
tldextract==5.1.3
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
typer==0.12.5
typing-extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
wheel==0.45.0
yarl==1.20.1
zxing-cpp==2.2.0
